Move the Package object to the DataModel namespace
