+++ Corgibytes.Freshli.Cli/Services/IAgentReader.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Corgibytes.Freshli.Cli.DataModel;
 using Corgibytes.Freshli.Cli.Functionality;
 using PackageUrl;
